+++ wpca/wpca.py
@@ -61,7 +61,11 @@ class WPCA(BaseEstimator, TransformerMixin):
         self.copy_data = copy_data
 
     def _center_and_weight(self, X, weights, fit_mean=False):
-        """Compute centered and weighted version of X.
+        """Compute centered and weighted version of X and adjust weights.
+
+        Input weights are inverse variance and adjusted weights are
+        inverse sigmas. Will produce a RuntimeWarning if any input
+        weights are negative.
 
         If fit_mean is True, then also save the mean to self.mean_
         """
@@ -75,6 +79,9 @@ class WPCA(BaseEstimator, TransformerMixin):
         X -= self.mean_
 
         if weights is not None:
+            # Convert from inverse variance to inverse sigmas.
+            # See eqn. 7 of Delchambre 2015 and issue #2.
+            weights = np.sqrt(weights)
             X *= weights
         else:
             weights = np.ones_like(X)
@@ -92,7 +99,7 @@ class WPCA(BaseEstimator, TransformerMixin):
 
         weights: array-like, shape (n_samples, n_features)
             Non-negative weights encoding the reliability of each measurement.
-            Equivalent to the inverse of the Gaussian errorbar.
+            Equivalent to the inverse variance when errors are Gaussian.
 
         Returns
         -------
@@ -127,6 +134,8 @@ class WPCA(BaseEstimator, TransformerMixin):
         self.explained_variance_ = evals[::-1]
         self.explained_variance_ratio_ = evals[::-1] / covar.trace()
 
+        self.n_iter_ = 1 # needed by sklearn.utils.estimator_checks
+
     def transform(self, X, weights=None):
         """Apply dimensionality reduction on X.
 
@@ -141,7 +150,7 @@ class WPCA(BaseEstimator, TransformerMixin):
 
         weights: array-like, shape (n_samples, n_features)
             Non-negative weights encoding the reliability of each measurement.
-            Equivalent to the inverse of the Gaussian errorbar.
+            Equivalent to the inverse variance when errors are Gaussian.
 
         Returns
         -------
@@ -174,7 +183,7 @@ class WPCA(BaseEstimator, TransformerMixin):
 
         weights: array-like, shape (n_samples, n_features)
             Non-negative weights encoding the reliability of each measurement.
-            Equivalent to the inverse of the Gaussian errorbar.
+            Equivalent to the inverse variance when errors are Gaussian.
 
         Returns
         -------
@@ -213,7 +222,7 @@ class WPCA(BaseEstimator, TransformerMixin):
 
         weights: array-like, shape (n_samples, n_features)
             Non-negative weights encoding the reliability of each measurement.
-            Equivalent to the inverse of the Gaussian errorbar.
+            Equivalent to the inverse variance when errors are Gaussian.
 
         Returns
         -------
@@ -235,7 +244,7 @@ class WPCA(BaseEstimator, TransformerMixin):
 
         weights: array-like, shape (n_samples, n_features)
             Non-negative weights encoding the reliability of each measurement.
-            Equivalent to the inverse of the Gaussian errorbar.
+            Equivalent to the inverse variance when errors are Gaussian.
 
         Returns
         -------