+++ CHANGES.md
@@ -4,3 +4,4 @@
 
 - Add ``copy_data`` argument in ``WPCA``. If set to False, this will prevent
   unnecessary copies of the data.
+- Use inverse variance consistently as the weight for ``WPCA`` and ``EMPCA``. Fixes issue #2.